move
org.camunda.bpm.engine.impl.telemetry.reporter.DiagnosticsCollector ->
org.camunda.bpm.engine.impl.telemetry.DiagnosticsCollector

org.camunda.bpm.engine.impl.telemetry.JavaClases ->
org.camunda.bpm.engine.impl.diagnostics
