+++ engine/src/main/java/org/camunda/bpm/engine/impl/cmd/GetTelemetryDataCmd.java
@@ -19,10 +19,10 @@ package org.camunda.bpm.engine.impl.cmd;
 import org.camunda.bpm.engine.impl.ProcessEngineLogger;
 import org.camunda.bpm.engine.impl.cfg.CommandChecker;
 import org.camunda.bpm.engine.impl.cfg.ProcessEngineConfigurationImpl;
+import org.camunda.bpm.engine.impl.diagnostics.DiagnosticsCollector;
 import org.camunda.bpm.engine.impl.interceptor.Command;
 import org.camunda.bpm.engine.impl.interceptor.CommandContext;
 import org.camunda.bpm.engine.impl.telemetry.dto.TelemetryDataImpl;
-import org.camunda.bpm.engine.impl.telemetry.reporter.DiagnosticsCollector;
 
 public class GetTelemetryDataCmd implements Command<TelemetryDataImpl> {
