+++ engine/src/main/java/org/camunda/bpm/engine/impl/diagnostics/CommandCounter.java
@@ -14,25 +14,38 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.camunda.bpm.engine.impl.telemetry;
+package org.camunda.bpm.engine.impl.diagnostics;
 
-import org.camunda.bpm.engine.impl.telemetry.dto.ApplicationServerImpl;
+import java.util.concurrent.atomic.AtomicLong;
 
-public class PlatformDiagnosticsRegistry {
+public class CommandCounter {
 
-  protected static ApplicationServerImpl applicationServer;
+  protected String name;
+  protected AtomicLong count = new AtomicLong(0);
 
-  public static synchronized ApplicationServerImpl getApplicationServer() {
-    return applicationServer;
+  public CommandCounter(String name) {
+    this.name = name;
   }
 
-  public static synchronized void setApplicationServer(String applicationServerVersion) {
-    if (applicationServer == null) {
-      applicationServer = new ApplicationServerImpl(applicationServerVersion);
-    }
+  public String getName() {
+    return name;
   }
 
-  public static synchronized void clear() {
-    applicationServer = null;
+  public void setName(String name) {
+    this.name = name;
   }
+
+
+  public void mark() {
+    count.incrementAndGet();
+  }
+
+  public void mark(long times) {
+    count.addAndGet(times);
+  }
+
+  public long get() {
+    return count.get();
+  }
+
 }
