+++ engine/src/main/java/org/camunda/bpm/engine/impl/diagnostics/DiagnosticsRegistry.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.camunda.bpm.engine.impl.telemetry;
+package org.camunda.bpm.engine.impl.diagnostics;
 
 import java.util.HashMap;
 import java.util.HashSet;
